% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stan_caterpillar.R
\name{stan_caterpillar}
\alias{stan_caterpillar}
\title{Create caterpillar plots from rstan's stanfit objects}
\usage{
stan_caterpillar(obj, pars, pars_labels = NULL, hpd = TRUE,
  order_medians = TRUE, horizontal = TRUE, alpha_bounds = 0.3)
}
\arguments{
\item{obj}{a \code{stanfit} object}

\item{pars}{scalar or vector regular expressions for paramater
labels that you would like to plot as declared in \code{model_code} from the
\code{\link{stan}} call.}

\item{pars_labels}{vector of parameter labels for the output plot. Important:
they must be in the same order as in the \code{stanfit} object when
\code{as.data.frame(obj)} is called.}

\item{hpd}{logical. If \code{TRUE} then the 90% and 95% highest probability
density intervals are found. If \code{FALSE} then the corresponding
central intervals are found.}

\item{order_medians}{logical. Whether or not to order the points by their
medians.}

\item{horizontal}{logical. Whether or not you would like the lines to be
horizontal}

\item{alpha_bounds}{numeric. Alpha transparency value for the uncertainty
bounds.}
}
\description{
Create caterpillar plots from rstan's stanfit objects
}
\details{
Points plot the simulations' medians, thin lines represent the 95%
HPD/central intervals, and thick lines represent the 90% HPD/central
intervals.
}
\examples{
\dontrun{
# Create Stan model
library(rstan)
scode <- "
   parameters {
       real y[2];
   }
   model {
       y[1] ~ normal(0, 1);
       y[2] ~ double_exponential(0, 2);
   }
"

# Run
fit <- stan(model_code = scode, iter = 10000, verbose = FALSE)

# Plot y[1] and y[2] parameters
stan_caterpillar(fit, pars = 'y\\\\[.*\\\\]')
}
}
\seealso{
\link{rstan}, \code{\link{stan}}, \code{ggmcmc}
}

