% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stan_caterpillar.R
\name{stan_catterpillar}
\alias{stan_catterpillar}
\title{A function to create caterpillar plots from rstan's stanfit objects}
\usage{
stan_catterpillar(obj, pars, pars_labels = NULL, hpd = TRUE,
  order_medians = TRUE, horizontal = TRUE)
}
\arguments{
\item{obj}{a \code{stanfit} object}

\item{pars}{character string, vector, or regular expression of paramater
labels that you would like to plot as declared in \code{model_code} from the
\code{\link{stan}} call.}

\item{pars_labels}{vector of parameter labels for the output plot. Important:
they must be in the same order as in the \code{stanfit} object when
\code{as.data.frame(obj)} is called.}

\item{hpd}{logical. If \code{TRUE} then the 90% and 95% highest probability
density intervals are found. If \code{FALSE} then the corresponding
central intervals are found.}

\item{order_medians}{logical. Whether or not to order the points by their
medians.}

\item{horizontal}{logical. Whether or not you would like the lines to be
horizontal}
}
\description{
A function to create caterpillar plots from rstan's stanfit objects
}
\details{
Points plot the simulations' medians, thin lines represent the 95%
HPD/central intervals, and thick lines represent the 90% HPD/central
intervals.
}
\seealso{
\link{rstan}, \code{\link{stan}}, \code{ggmcmc}
}

